update the build file
