+++ package.json
@@ -2,7 +2,7 @@
   "name": "@spot-flow/checkout-inline-js",
   "version": "0.0.4",
   "description": "",
-  "main": "dist/checkout-inline.js",
+  "main": "dist/checkout-inline-js.js",
   "source": "src/index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
