+++ vite.config.ts
@@ -13,8 +13,6 @@ export default defineConfig({
       external: [],
       output: {
         globals: {},
-        format: "umd",
-        exports: "named",
       },
     },
   },
